moving archive channel handler to separate file
